pyproject.toml
[tool.poetry]
name = "pythondaq"
version = "0.1.0"
description = ""
authors = ["JukkaJohn <jukkajohn2002@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
PyVISA-py = "^0.5.2"
pyserial = "^3.5"
numpy = "^1.21.4"
matplotlib = "^3.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
diode = 'pythondaq.views.diode:cmd_group'
